+++ CHANGES.rst
@@ -31,6 +31,9 @@ Changelog
 - Added due date default value configuration
   [bleybaert]
 
+- Added batch actions buttons (labels, senders, recipients)
+  [sgeulette]
+
 2.1 (2018-08-22)
 ----------------